This commit makes the information about the venue, submiting talks and the about the conference more visible
To this commit make sense someone need to remove the about post
from the blog, since it has repeated information

Move the about with the information of the venue to the top.
After information about the theme and about the talks,
including the warning about the date and the link to submit it
Add a Planning your attendance header where the attendance
information are.

Remove most of repeated glasgow and perl and about words
as this hurt google indexes